+++ packages/terra-dev-site/CHANGELOG.md
@@ -1,6 +1,8 @@
 # Changelog
 
 ## Unreleased
+* Changed
+  * Upgraded to intl v5
 
 ## 7.6.0 - (February 14, 2022)