+++ packages/terra-dev-site/CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## Unreleased
 
+* Breaking
+  * Dropped v2-v4 react-intl support
+
 ## 7.6.0 - (February 14, 2022)
 
 * Changed